exams add

+++ Backend/src/main/java/com/example/surveillance/repositories/MatiereRepositoriy.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface MatiereRepositoriy extends JpaRepository<Matiere,Integer> {
+public interface MatiereRepositoriy extends JpaRepository<Matiere,Long> {
 
 }
